rename 'MetamodelUtility' to 'MetaModelUtils'
