+++ packages/legend-studio/src/components/editor/edit-panel/mapping-editor/NewMappingElementModal.tsx
@@ -24,7 +24,7 @@ import {
   createFilter,
 } from '@finos/legend-studio-components';
 import type { SelectComponent } from '@finos/legend-studio-components';
-import { fromElementPathToMappingElementId } from '../../../../models/MetaModelUtility';
+import { fromElementPathToMappingElementId } from '../../../../models/MetaModelUtils';
 import { MappingEditorState } from '../../../../stores/editor-state/element-editor-state/mapping/MappingEditorState';
 import {
   UnsupportedOperationError,
