+++ packages/legend-studio/src/models/metamodels/pure/model/packageableElements/domain/ConcreteFunctionDefinition.ts
@@ -21,7 +21,7 @@ import {
   deleteEntry,
   changeEntry,
 } from '@finos/legend-studio-shared';
-import { hashLambda } from '../../../../../MetaModelUtility';
+import { hashLambda } from '../../../../../MetaModelUtils';
 import type { Hashable } from '@finos/legend-studio-shared';
 import { CORE_HASH_STRUCTURE } from '../../../../../MetaModelConst';
 import type { PackageableElementVisitor } from '../../../model/packageableElements/PackageableElement';
