+++ packages/legend-studio/src/models/protocols/pure/v1/V1_PureGraphManager.ts
@@ -43,7 +43,7 @@ import {
   GraphError,
   SystemGraphProcessingError,
   DependencyGraphProcessingError,
-} from '../../../MetaModelUtility';
+} from '../../../MetaModelUtils';
 import type { AbstractEngineConfig } from '../../../metamodels/pure/action/AbstractEngineConfiguration';
 import type {
   EngineSetupConfig,
