+++ packages/legend-studio/src/models/protocols/pure/v1/model/packageableElements/domain/V1_DerivedProperty.ts
@@ -15,7 +15,7 @@
  */
 
 import { hashArray } from '@finos/legend-studio-shared';
-import { hashLambda } from '../../../../../../MetaModelUtility';
+import { hashLambda } from '../../../../../../MetaModelUtils';
 import type { Hashable } from '@finos/legend-studio-shared';
 import { CORE_HASH_STRUCTURE } from '../../../../../../MetaModelConst';
 import type { V1_Multiplicity } from './V1_Multiplicity';
