+++ packages/legend-studio/src/models/protocols/pure/v1/model/packageableElements/function/V1_ConcreteFunctionDefinition.ts
@@ -15,7 +15,7 @@
  */
 
 import { hashArray } from '@finos/legend-studio-shared';
-import { hashLambda } from '../../../../../../MetaModelUtility';
+import { hashLambda } from '../../../../../../MetaModelUtils';
 import { CORE_HASH_STRUCTURE } from '../../../../../../MetaModelConst';
 import type { V1_RawVariable } from '../../../model/rawValueSpecification/V1_RawVariable';
 import type { V1_Multiplicity } from '../../../model/packageableElements/domain/V1_Multiplicity';
