+++ packages/legend-studio/src/models/protocols/pure/v1/model/packageableElements/store/relational/mapping/V1_RelationalPropertyMapping.ts
@@ -19,7 +19,7 @@ import { CORE_HASH_STRUCTURE } from '../../../../../../../../MetaModelConst';
 import type { Hashable } from '@finos/legend-studio-shared';
 import type { V1_PropertyMappingVisitor } from '../../../../../model/packageableElements/mapping/V1_PropertyMapping';
 import { V1_PropertyMapping } from '../../../../../model/packageableElements/mapping/V1_PropertyMapping';
-import { hashObjectWithoutSourceInformation } from '../../../../../../../../MetaModelUtility';
+import { hashObjectWithoutSourceInformation } from '../../../../../../../../MetaModelUtils';
 import type { V1_RawRelationalOperationElement } from '../model/V1_RawRelationalOperationElement';
 
 export class V1_RelationalPropertyMapping
