+++ packages/legend-studio/src/models/protocols/pure/v1/transformation/pureGraph/to/helpers/V1_ValueSpecificationBuilderHelper.ts
@@ -25,7 +25,7 @@ import {
   TYPICAL_MULTIPLICITY_TYPE,
   PRIMITIVE_TYPE,
 } from '../../../../../../../MetaModelConst';
-import { GraphError } from '../../../../../../../MetaModelUtility';
+import { GraphError } from '../../../../../../../MetaModelUtils';
 import {
   LambdaFunction,
   FunctionType,
