+++ packages/legend-studio/src/stores/ViewerStore.tsx
@@ -31,7 +31,7 @@ import {
 } from '@finos/legend-studio-shared';
 import { Workspace } from '../models/sdlc/models/workspace/Workspace';
 import type { Entity } from '../models/sdlc/models/entity/Entity';
-import { GraphError } from '../models/MetaModelUtility';
+import { GraphError } from '../models/MetaModelUtils';
 import { useLocalObservable } from 'mobx-react-lite';
 import { EDITOR_MODE, TAB_SIZE } from './EditorConfig';
 import type { ViewerRouteParams } from './Router';
